Update MidtermProjectYAML.yml
To use this you must change on line 366 where you are defining the bucket policy. You must change the file path at the end with 9 numbers to the account number of the account in which you are using. Also do not use this if you do not want to lose money. It builds an RDS DB. Also make sure the bucket name around line 350 is unique.